chore(deps): upgrade hapi, hapi-auth-jwt2 and typescript
BREAKING CHANGE: This version of Prism is now only compatible with the
latest version of the Hapi API. Although the API of Prism itself has not
changed, it now only work with a suitable version of Hapi.

BREAKING CHANGE: This version of Prism is now compiled using TypeScript
3.x; projects written in TypeScript that use Prism may require to use a
matching version of TypeScript to ensure correct type compatibility.

BREAKING CHANGE: The Security plugin in this version of Prism now
uses the latest version of the `hapi-auth-jwt2` plugin.